clients/python-fastapi/generated/tests/test_characteristic_api.py
# coding: utf-8

from fastapi.testclient import TestClient




def test_characteristic_list(client: TestClient):
    """Test case for characteristic_list

    
    """
    params = [("limit", 56),     ("offset", 56)]
    headers = {
    }
    # uncomment below to make a request
    #response = client.request(
    #    "GET",
    #    "/api/v2/characteristic/",
    #    headers=headers,
    #    params=params,
    #)

    # uncomment below to assert the status code of the HTTP response
    #assert response.status_code == 200


def test_characteristic_read(client: TestClient):
    """Test case for characteristic_read

    
    """

    headers = {
    }
    # uncomment below to make a request
    #response = client.request(
    #    "GET",
    #    "/api/v2/characteristic/{id}/".format(id=56),
    #    headers=headers,
    #)

    # uncomment below to assert the status code of the HTTP response
    #assert response.status_code == 200

